feat: remove -d option, cli rename into cssobj; test: folder process, change test format

+++ test/test.js
@@ -8,10 +8,12 @@ function formatResult (str) {
   return str.replace(/\n/g, '\\n').replace(/\s+$/, '[ ]')
 }
 
-function testCli(testFile, option, targetFile, done) {
+function testCli(source, option, target, done) {
+  var fileRe = /^file::/
+  var folderRe = /^folder::/
   var cli = 'cli/cssobj-converter.js'
   // var format = path.extname(testFile).slice(1).toLowerCase()
-  var cliProcess = spawn('node', [cli].concat(option).concat([testFile]))
+  var cliProcess = spawn('node', [cli].concat(option).concat([source]))
   var srcFile = option[0]=='-o' && option[1]
 
   var output = '', msg=''
@@ -24,29 +26,40 @@ function testCli(testFile, option, targetFile, done) {
   cliProcess.on('exit', function() {
     if(msg) return done(msg)
 
-
-    if(srcFile){
-      if(output) return done('should have empty output, but the output is:\n'+ output)
+    // file type result check
+    if(fileRe.test(target)) {
       try {
-        output = fs.readFileSync(srcFile, 'utf8')
+        target = fs.readFileSync(target.replace(fileRe, ''), 'utf8')
       } catch(e) {
         return done(e)
       }
-      if(srcFile) fs.unlinkSync(srcFile)
+      if(srcFile) {
+        if(output) return done('should have empty output, but the output is:\n'+ output)
+        try {
+          output = fs.readFileSync(srcFile, 'utf8')
+        } catch(e) {
+          return done(e)
+        }
+        if(srcFile) fs.unlinkSync(srcFile)
+      }
+      expect(output).equal(target)
     }
 
-    try {
-      var target = fs.readFileSync(targetFile, 'utf8')
-    } catch(e) {
-      return done(e)
+    // folder type check
+    if(folderRe.test(target)) {
+      var result = require('./'+path.join(target.replace(folderRe, '')))
+      fs.readdirSync(source, 'utf8').forEach(function(file) {
+        if(/\.js$/.test(file)) {
+          if(!(file in result)) return done(file + ' not exists in result json.')
+          var filePath = path.join(source, file)
+          var code = fs.readFileSync(filePath, 'utf8').trim()
+          fs.unlinkSync(filePath)
+          expect(result[file]).equal(code)
+        }
+      })
     }
-
-    var result = '', line=0
-
-    expect(output).equal(target)
-
-    done(result)
-
+    // ensure test done
+    done()
   })
 }
 
@@ -55,37 +68,43 @@ describe('Test cli converter', function () {
   this.timeout(5000)
 
   it('with pretty false', function (done) {
-    testCli('test/cli/test.css', ['-p', false], 'test/cli/test.js', done)
+    testCli('test/cli/test.css', ['-p', false], 'file::test/cli/test.js', done)
   })
 
   it('with pretty true', function (done) {
-    testCli('test/cli/test.css', ['-p', true], 'test/cli/test-pretty.js', done)
+    testCli('test/cli/test.css', ['-p', true], 'file::test/cli/test-pretty.js', done)
   })
 
   it('with output file', function(done) {
-    testCli('test/cli/test.css', ['-o', 'cli-temp.js'], 'test/cli/test-pretty.js', done)
+    testCli('test/cli/test.css', ['-o', 'cli-temp.js'], 'file::test/cli/test-pretty.js', done)
   })
 
   it('format less no -f', function(done) {
     // without -f option
-    testCli('test/cli/test.less', [], 'test/cli/test-less.js', done)
+    testCli('test/cli/test.less', [], 'file::test/cli/test-less.js', done)
 
   })
 
   it('format less with -f', function(done) {
     // with explicitly -f less option
-    testCli('test/cli/test.less', ['-f', 'less'], 'test/cli/test-less.js', done)
+    testCli('test/cli/test.less', ['-f', 'less'], 'file::test/cli/test-less.js', done)
 
   })
 
   it('format with -c', function(done) {
     // with explicitly -f less option
-    testCli('', ['-c', 'p{color: red;}'], 'test/cli/command-line.js', done)
+    testCli('', ['-c', 'p{color: red;}'], 'file::test/cli/command-line.js', done)
 
   })
 
   it('with output file and no pretty', function(done) {
-    testCli('test/cli/test.css', ['-o', 'cli-temp.js', '-p', false], 'test/cli/test.js', done)
+    testCli('test/cli/test.css', ['-o', 'cli-temp.js', '-p', false], 'file::test/cli/test.js', done)
   })
 
+})
+
+describe('test with folder', function() {
+  it('should accept folder', function(done) {
+    testCli('test/cli/folder', ['-p', false], `folder::cli/folder.js`, done)
+  })
 })
